refactor: authz permission annotation, handler
